Move inner classes out and organise in format package
